this is a bit better, but still kinda bad
