+++ evals/elsuite/utils.py
@@ -13,11 +13,6 @@ from evals.prompt.base import (
     is_chat_prompt,
 )
 
-from langchain.text_splitter import RecursiveCharacterTextSplitter
-from langchain.embeddings.openai import OpenAIEmbeddings
-from langchain.docstore.document import Document
-from langchain.vectorstores import FAISS
-
 
 def get_answer(text, answer_prompt, ignore_case=False):
     if ignore_case:
@@ -145,48 +140,6 @@ def format_prompt(
         prompt = format_necessary(prompt, allow_missing=allow_missing, **kwargs)
     return prompt
 
-
-def process_text(text):
-    docs = text_to_docs(text)
-    embeddings = OpenAIEmbeddings()
-    db = FAISS.from_documents(docs, embeddings)
-    return db
-
-def retrieve_context_from_db(query, db, k=3):
-    documents = db.similarity_search(query=query, k=3)
-    return ''.join([doc.page_content for doc in documents])
-
-def text_to_docs(text):
-    """Converts a string or list of strings to a list of Documents
-    with metadata."""
-    if isinstance(text, str):
-        # Take a single string as one page
-        text = [text]
-    page_docs = [Document(page_content=page) for page in text]
-
-    # Add page numbers as metadata
-    for i, doc in enumerate(page_docs):
-        doc.metadata["page"] = i + 1
-
-    # Split pages into chunks
-    doc_chunks = []
-
-    for doc in page_docs:
-        text_splitter = RecursiveCharacterTextSplitter(
-            chunk_size=2000,
-            separators=["\n\n", "\n", ".", "!", "?", ",", " ", ""],
-            chunk_overlap=0,
-        )
-        chunks = text_splitter.split_text(doc.page_content)
-        for i, chunk in enumerate(chunks):
-            doc = Document(
-                page_content=chunk, metadata={"page": doc.metadata["page"], "chunk": i}
-            )
-            # Add sources a metadata
-            doc.metadata["source"] = f"{doc.metadata['page']}-{doc.metadata['chunk']}"
-            doc_chunks.append(doc)
-    print("Length of doc_chunks: ", len(doc_chunks))
-    return doc_chunks
 
 
 class PromptFn:
